semantic_segmentation/fn-sseg/requirements.txt
expiringdict
matplotlib
numpy
Pillow
sagemaker>=2,<3
